backend/requirements.txt
anthropic>=0.39
openpyxl>=3.1
python-docx>=1.1
pdfplumber>=0.10
Pillow>=10.0
pandas>=2.0
